Fix missing newline problem with ast.
When the ast module parses a docstring, it strips the whitespace from
the end.  We expected to see whitespace after the end of the sections
(such as Args or Returns).  This will probably make it slightly more
difficult to verify the general formatting of a google docstring in this
regard.

Also, applied darglint to all of its own files.