Wrap error gathering into a class.
The IntegrityChecker class has a method, run_checks(), which accepts an
ast.FunctionDef.  The method aggregates errors in the class. Once all
the functions in a module have been passed through it, you can call its
get_error_report() method, which returns a string representation of all
the errors.

In the future, the get_errors_report() method will take an optional
parameter to specify the granularity of the message.  It will pass that
option into the appropriate DarglintError so that they can render
themselves.