Add checks for a Raises section.
Currently, the checks will warn for missing or excess exceptions or
errors raised in the "Raises"  section.  However, excess exceptions
will be disabled by default in the future.  That will be to allow
code such as

    def quotient(x, y):
        """Get the quotient of the two numbers.

        Args:
            x: The numerator.
            y: The denominator.

        Raises:
            ZeroDivisionError: If y is zero.

        Returns:
            The quotient of the two numbers.

        """
        return x / y

Here, even though a ZeroDivisionError isn't explicitly raised, we may
still want to warn the user in the interface.  Of course, a way of
getting around this would be to have the body look like

    if y == 0:
        raise ZeroDivisionError
    return x / y

or using a try/except block.  But, that may be too much of an
inconvenience for some.