Add type checks.
The docstring parser can now parse types from the docstring.
FunctionDescriptions now also record types.  Finally, the
IntegrityChecker was updated to check parameter types in the
docstring against type hints.

This is probably of very limited utility.  If you're using type
hints, you really don't need to document the types in the docstring.
However, if you do mix them, it's an easy for others to be confused.