Add config file in current directory.
Darglint now will allow you to ignore errors using a configuration file
in the current directory (wherever you are calling darglint from.)  The
config file can have the name ".darglint", or "setup.cfg", or "tox.ini".
Darglint will choose the first file with a valid name that it
encounters, checking in that order.

The syntax for the config file should be:

    [darglint]
    ignore=I101,I102

If we wanted to ignore all instances of I101 and I102.