Handle poorly formatted config files.
If a file matches one of the default filenames for configuration files,
and the configparser raises a ParsingError, then a warning message is
logged out to error using the default logging settings for python.
That file will then be skipped and the search will continue until a
valid file is found, or until there are no more files/directories to
check.