Add GenericSyntaxError, S001; add optional errors.
GenericSyntaxError will, for the moment, capture any ParserExceptions
raised in the parsing of a docstring.  This will be able to handle the
majority of malformed docstrings.

Also, optionally raising exceptions on errors is possible.  For example,
if we were to invoke darglint with the following option:

    darglint --raise-syntax html_linter.py

Then syntax errors will be propagated.  This will make it slightly
easier to develop darglint itself, since it will allow us to see the
stack trace of the exception. (And so, we can easily see what part of
the docstring is raising the exception.)