Add function to get logger; skip bare raises.
If a function raises the same exception which it catches, then it is
simply (for the time being) skipped in the checks.  Really, we should
check the context of the error-catching to see if there is a type.  Of
course, even then, it wouldn't necessarily be the most specific type.

For example, if we have a function such as:

    def do_something_dangerous():
        try:
            dangerous_action()
        except Exception as ex:
            raise

This would have previously raised an error and darglint would have
crashed.  Now, however, it is simply skipped.  We still might want to
see if there are any errors described in the docstring, though.