Add check that description comes after arg.
After each argument or description, there should be a description
of some kind.  Previously, this would kill the parser (as it was
assumming content after the colon.  Now, it raises a ParserException.

For example, the following function will now raise a ParserException:

    def xsum(*items):
        """Return the sum of the arguments.

        Args:
	    *items:

	Return:
	    The sum of the arguments.

	"""
	return sum(items)

As this raises a ParserException, it can only be ignored globally
(through the config file) by excluding S001, (GenericSyntaxError).
However, that's too broad of an item to exclude.  But, thankfully,
to exclude it in the current line, you can use a noqa statement:

    def xsum(*items):
        """Return the sum of the arguments.

        Args:
	    *items: # noqa: S001

	Return:
	    The sum of the arguments.

	"""
	return sum(items)

This actually just reads "# noqa: S001" as the description, but it
fulfills the intent of the author. (If they really want to keep others
in the dark.)