Add parsing item for composite section.
This required increasing lookahead to 3: in order to see that the next
line is not the start of a definition (instead of another line in the
current definition), we have to look three tokens ahead.

The Peaker class was made to have any constant amount of lookup.  This
means that the parser will have to have a lookahead of three, but the lexer
can retain a lookahead of one.