Add first version of new parser.
This parser builds an AST which can be traversed and reasoned over by
the IntegrityChecker.

Example Usage:

    parse(Peaker(lex('Return pi.')))

Which would parse the short docstring, 'Return pi.'.

There are still many tests to run and several things to add to this
parser.  For example, it does not currently parse noqa statements.
Additionally, it does not pass line numbers onto nodes (or reference
them in error messages.)  Finally, it has not yet been integrated into
Darglint as a whole.