Work in old examples.
For the moment, removed the ability to have a simple section start on
the same line as the keyword.  For example, previously we were allowed
to have things like

    def get_x():
        """Get the value.

        Returns: The value.

        """
        return 9

Now, however, it would not be allowed.  We would have to have

    def get_x():
        """Get the value.

        Returns:
            The value.

        """
        return 9

This is a pretty minor change, but it may be added back in later.