Update line numbers to be closer to source.
Line numbers are now closer to the actual source of issues.
If the error describes a missing item, the error will default to
the top of the docstring.

Line numbers initially tracked in tokens.  From there, they are passed
into Nodes, and from Nodes into Errors.  It could possibly be simpler to
maintain line numbers in global variables while parsing, but this might
give us more flexibility down the road.  (For example, if we move to
using multithreading to parse docstrings, it will make maintaining state
easier, since nearly everything is referentially transparent.)