Add line/short-description parsing of Sphinx docs.
This commit is the first in an effort to include Sphinx-style docstrings
in darglint.  Some options will be slightly more strict than
Sphinx-style docstrings, which are relatively permissive.  For example,
darglint will expect that there is a short description.  In the future,
these might become optional with a flag.

Currently, there are a lot of similar tokens between Google-Style and
Sphinx-Style docstrings.  For example, all of the terminal tokens are
the same.  Furthermore, many of the helper functions for parsing are
already being reused.   It will make sense, going forward, to
restructure the project to look like:

    darglint/
    ...
    darglint/parse/
    darglint/parse/__init__.py
    darglint/parse/common.py
    darglint/parse/google.py
    darglint/parse/sphinx.py
    ...

That way, we will be able to do something like:

    # In darglint/parse/sphinx.py

    from .utils import (
        Assert,
        p_is,
        ...
    )