Add parse item for sphinx, dot renderer.
The dot renderer, in darglint.utils, takes an AST and renders a dot file.
The generated dot file can then be rendered using graphviz.  This is
useful for debugging darglint.

Example Usage:

    from darglint.parse.google import *
    from darglint.lex import *
    from darglint.utils import generate_dot

    docstring = '''Add two numbers.

        Args:
            x: The first number to add.
            y: The second number to add.

        Returns:
            The sum of the two numbers.

    '''
    node = parse(Peaker(lex(docstring), lookahead=3))
    with open('~/Documents/tmp/docstring.dot', 'w') as fout:
        fout.write(generate_dot(node))

A script to render the first docstring found in a file will be added to
this utility. It may eventually be added into the command line utility
if its use in debugging is significant.