Add method to consolidate AST for sphinx docstring
This forces the sphinx AST to be structured the same as the Google AST.
In the future, we could add a similar method for Google-style docstrings
to translate them to Sphinx format.  With Node.reconstruct_string(),
that would allow us to automatically translate between docstring styles.
Of course, we would have to change how keywords/etc. are rendered. (They
are currently reconstructed using their value, but we would have to
change it to depend on the node type.)