Add Sphinx Docstring object and configuration.
In order to accommodate the different trees produced by a Sphinx
docstring and a Google docstring, I've abandoned the AST consolidation
approach in favor of abstracting out an interface for Docstring and
adding a Docstring factory.

Actually transforming the AST, while it would have had nice benefits,
wouldn't have saved any code or time.  The transformation functions
turned out to be very ugly, and required several traversals of the tree
to accomplish.  In addition, it wouldn't have made Darglint more
flexible in case we ever want to add a new docstring format. (Or edit
the existing ones.)

Instead, I've created a BaseDocstring class, which represents the
interface for Docstrings.  Subclasses of BaseDocstring should know how
to return the requested data from the AST for that docstring style.
Currently, it looks like this:

                  +----------------------+
                  |     BaseDocstring    |
                  |    << Interface >>   |
                  +======================+
                  | get_argument_types() |
                  | ...                  |
                  +----------------------+
                     ^                ^
                    /_\              /_\
                     |                |
   +----------------------+     +----------------------+
   |   sphinx.Docstring   |     |   google.Docstring   |
   +----------------------+     +----------------------+
          .:.                                  .:.
           :                                    :
      << create >>                         << create >>
           :                                    :
           :                                    :
           :     +-----------------------+      :
           :     |       Docstring       |      :
           :     +=======================+      :
           ......|.from_sphinx()         |      :
                 |                       |      :
                 | from_google().........|.......
                 +-----------------------+

This will probably be revised so that it makes a little more sense.
But, the bottom Docstring (in darglint/docstring/docstring.py) is a
Factory for instances of BaseDocstring.  Currently, the instances
are created using the methods from_sphinx() and from_google().  This
could be changed to a single method that takes the Configuration
instance.

In any case, adding a new docstring format will be as easy as creating
a new parser and a new BaseDocstring subclass, and updating the
Factory.  The IntegrityChecker, then, doesn't have to worry about which
format, and just performs checks based on the BaseDocstring interface.