Add inline sphinx types for parameters/etc.
Users can write a docstring which contains the following field:

    :param int foo: ...

In which *int* refers to the type of the parameter.  This is now
a possibility.

A check should be added for different types of fields.  For example,

    :returns int: ...

Should be possible, but

    :returns int shaz: ...

Should not.  Probably the parse_item_name function should delegate to
different parsers based on the field type.