Add excess variable check for Sphinx.
Sphinx has a item of the type :var:, which describes variables.
While you definitely don't want to describe all variables in a function,
you probably also don't want to describe variables which you've removed
or renamed.  So, now there is the ExcessVariableError.

For example,

    def cookie_area(width):
        """Calculate the area of the cookie.

        :param width: The width of the cookie.
        :var pi: The value of PI.
        :return: The cookie's area.

        """
        radius = width / 2
        return 3.1415 * radius**2

Will raise an exception because pi is not defined.  However, no
exception will be raised because "radius" was not described.