Intepret whitespace-only lines as newlines.
This really should only affect lines immediately after sections which
have the same or greater level of indentation (but only whitespace) as
we would expect to occur in the described items of those sections.

For example, if we have

    "Args:\n    x: The first argument\n    \nReturns:\n    Nothing\n\n"
                                           ^
                                           |
                                    Expects a word + colon here.

Darglint would have failed to parse the section at the indicated
location.

Now, it treats it as a newline.  The unfortunate side effect of this
choice is that blocks of text within arguments will not be parsed as a
part of that section.  For example, if we have

    Args:
        x: A value which satisfies the following constraint:

            x^-1 > 2

        ignore: Ignore any errors.

Then we will always get a missing parameter exception for the parameter,
ignore.  Since this sort of format leads to pretty ugly
arguments/raises/etc. sections, it's probably not a big deal.
Eventually it might get handled.

Resolves: Issue #7.