Fix type errors in parse/common.
Unfortunately, to appease the type checker, we have to have peak
actually return a value, or add an assertion.  (Our custom Assert and
AssertNotEmpty functions do not preclude type errors.)

To solve this, a different, less safe, peak function was added to the
peaker, rpeak.  It will raise an exception if there are items present.
Alternatively, we could change the custom assertion functions to the
built-in assert.  Or, we could add explicit checks, or use
short-circuiting.  (For example, we could have an invalid token and do
something like

    (peaker.peak() or invalid).value

Which would raise an exception.  Ultimately, this is about the same.