Add initial test cases for new parser.
The new parser is going to have to handle an ambiguous grammar.  So, it
is going to be something like a GLR or Chart parser.

This will make parsing quite a bit more complex that it is currently.
One way to simplify it would be to break the stylistic warnings and hard
errors into two steps. (This is kind of already done, with introspection
on the resulting tree of the parse.)