Add basic CYK implementation.
This implementation of a CYK parser is still pretty basic.  It doesn't,
as of yet, handle probabilistic grammars.  It could also be cleaned up.
In particular, an interface for Grammars will be created, one which can
be nested.

One problem with this CYK parser is that it is O(n^3).  Running it
against even relatively short strings can result in a large runtime.
For example, running it against the "SmallGrammar" in
"tests/test_cyk.py" showed that at around 200 symbols, the parse
function takes about four seconds.

This can be alleviated in several ways.  First, we can reduce the
parsing of docstrings to subproblems.  That is, each individual section
can have its own parser.  That would involve nesting several CYK
parsers, which wouldn't seem to get us very far.  However, if we thread
those subparsers, we can achieve a faster parse overall.  And since the
keywords which designate the start of each section won't come up very
often, we probably won't be calling these parsers multiple times.

Furthermore, we can condense the input.  Most of a docstring is
composed not of syntactically meaningful tokens, but of plain words.  We
can group words which do not match a keyword (we'll have to be careful
about this, to make sure we capture possible misspellings of
keywords.)  Once all of these are grouped, parsing a docstring should be
relatively quick.