Begin bnf_to_cnf conversion tool.
This tool will make it easier to write grammars by converting a BNF
grammar to a CNF grammer which can be consumed by the CYK parser.
The tool will only be used for development, so it is included in the bin
folder, and not used or packaged by default.

Ironically, this tool will need to parse the BNF grammar.  For that, I'm
using Lark parser (see https://github.com/lark-parser/lark).  For
darglint the parser needs to communicate line numbers and other context,
which is generally not easy to handle with parser generators.
For this application, however, a parser generator is fine: it's okay if
the grammar is a little painful to edit.