Add ε-rule elimination to bnf to cnf utility.
This eliminates any ε-rules which are in the grammar.  The current
implementation is very inefficient, which should be fine as this is just
an internal tool.  It could be sped up quite a bit, however.

It also maxes out at 100 iterations of ε-rule removal.  Most of the
grammars should be pretty small.  But, if there are still ε-rules after
the transformation, then the `max_iterations` number needs to be
increased, or there is some problem with the grammar itself (that
prevents the rules from being eliminated.)