Update CYK parser to take token types as terminals
This allows the output of the lexer to be used in the cyk parser.
One advantage of this is that transformations can be done on the tokens
from the lexer to make parsing faster.  For example, in a docstring, we
don't need to encode individual words, as was done previously: we can
combine runs of words into a single word.  This could greatly reduce the
number of tokens which need to be parsed, and so increase the overall
speed by a significant factor.

The plan had been to encode all non-terminal nodes as some subclass of a
`BaseNodeType` class.  However, that would not be possible: the
`bnf_to_cnf` tool has to be able to create new, arbitrarily named nodes.
Since enumerations aren't arbitrary, we can't entirely encode this node
type as an enum.  We could have a mix of enumerations and strings, since
the original enumerations would still be present.  However, this likely
isn't very helpful.

Instead, the next step will be to allow annotations in the BNF
description which translate to a specific error type during parsing.
This will allow us to catch all style errors during parsing, and all
semantic errors after reflecting over the tree.