Add annotations to bnf_to_cnf.
Now we can add annotations to the grammar.  One small caveat is
that it does not correctly translate annotations (or probably the
grammar itself) if you have multiple products with the same LHS.
So you have to use the `|` operator.

For example, the following will not be translated correctly:

    <A> ::= <B><C>
    <B> ::= "B"
    @Q <B> ::= "D"
    <C> ::= "C"

Instead, it should be

    <A> ::= <B><C>
    <B> ::= "B"
        | @Q "D"
    <C> ::= "C"

Which looks cleaner, anyway, so this probably won't be fixed.