Add annotation to python rendering.
Productions with annotations expect a static constructor,
`with_annotations` such as:

    P.with_annotations(
        [S001],
        'item_head',
        NodeType.WORD,
    )

Where the annotations could represent anything, but in this case
represent errors which should be reported if that particular production
is in use.  In the above example, a different production,

    P('item_head', ('word', 'colon'))

Would come first (and so receive priority.)  The error S001 would be
raised if, then, if there wasn't a colon to be parsed.