Add annotations to grammar object.
The next step will be to actually use those annotations in some way.
One option is to use an emitter in the CYK parse function.  However,
that introduces state in a non-referentially transparent way, which
may introduce race conditions later on if we use threading.

Another option is to simply pass the annotations on with the nodes.
Then, the tree will have to be walked and annotations will have to be
gathered as a separate step, which could make it a little slower.
Thankfully, the tree once parsed is quite small for a docstring, so it
may not be too bad.  This has the added advantage of preserving the
context of the annotations, which may be important for getting line
numbers.

A final option is to return a list of annotations along with the parsed
tree.  However, the annotations would have to be modified to hold their
line number.