Add annotations to parsed node.
When a node for an annotated productions is produced in the CYK table,
it receives those annotations.  In our case, those are error messages
which will be instantiated and aggregated for display to the user.
The CykNode will eventually know how to instantiate the error, because
it contains the context of the error's occurence.  (Line numbers, and
the strings contained in particular.)