Add basic outline for completing refactor.
Two decorators for tests were introduced, `replace` and `remove`.
Tests marked with `remove` should be removed by the end of the
refactor.  Most of these tests relied upon some implementation detail,
and were used to facilitate test-driven development. (Maybe those tests
should have been marked differently from the get-go?)  Others simply
need to be replaced to use the new parser.  These are marked with
`replace(name)` where name is the name of the method replacing them.

Once refactoring is completed, we can change the value of
`tests.utils.REFACTORING_COMPLETE` to True to see if there are any
methods which have not been removed or handled.  Then we would remove
the old methods, rename the appropriate files, and set the above global
back to False.

One issue already apparent is that when parsing the grammar,
non-terminal nodes will likely be removed when translating to CNF.  So,
we'll have to keep semantically meaningful nodes as low in the tree as
possible.  For example, if a returns section is being processed, then we
have to make sure we process the word "Returns" as the symbol,
"returns", and not "word".  Otherwise, we will know that the docstring
follows the correct format, but we won't be able to make meaningful
inferences over them.

Probably, we'll have to add some implementation-specific testing to
ensure that certain symbols are present in the grammar, so that we can
perform the correct inferences.  Likely, this will simplify the process
of checking docstrings, as we will only need to grab those terminals and
won't need to assume some structure in the tree.