Add short description, reconstruct_string.
The method `CykNode.reconstruct_string` will be used to rebuild
docstrings from their parsed structure.  This is useful for testing,
and may allow us to automatically fix stylistic errors in docstrings
in the future.

In order to fix stylistic errors, each error will have a method, `fix`,
which can be used to fix the nodes it describes.  When reconstructing
the string, we will go through the annotations and replace the string
we reconstructed with the result of their fix method.  (Or we will pass
the node itself to `fix` and allow it to construct a string.  This is
a less likely approach, though, because it would be difficult to fix
multiple problems with a node -- we would have to fix and reparse the
node several times.)

This will also allow us to set a strictness to the `fix` method.  We
will apply the `fix` method to any errors whose `strictness` is less
than (or equal to) the strictness passed to `fix`.  So, for example,
if we called the `fix` method with a strictness of 3, it will fix any
style errors at levels 1, 2, or 3.