Add compound sections to CYK grammar.
Adds the following items to the new parser for docstrings:

- Arguments section
- Raises section

Types still have to be handled in the items section, but otherwise most
aspects are present.

The parser has slowed down substantially: it takes nearly a second to
parse a moderate-sized docstring.  This should be able to be fixed later
on, though.  Probably, the size of the generated grammar is ballooning.
We can reduce the size of the generated grammar pretty easily.  Then we
can try inlining the grammar, if necessary. We can switch to use
multi-threading (although that won't fix the problem of the size of the
table, it will allow us to process more items concurrently, so it will
make parsing a whole project much faster).  And we could implement the
parsing algorithm and grammar in C (I'd rather not do this, though, as it
would require building darglint on the host.)