Retain start symbol.
Previously, the start symbol was a production, and so the symbols
it pointed to could be lost.  Now, it is a separate construct,
and will always retain the top symbol.  This will make analyzing the
resulting tree at least somewhat easier.

For example, in the grammar:

    start: <head>
    <head> ::= <args> <colon>
    # ...

The node, "head", will always be the first node if it was successfully
parsed.