Replace parser tests using old parser.
A couple of edge cases need to be handled yet.  In particular, it's
possible that sections get separated by an indented line (this was an
issue at one point.)  In this case, we should still separate them,
and issue a style warning.

However, this involves some custom logic with the top-level parser,
since that is what is being used to separate the sections.