Add terminal newlines to sections.
The end of the docstring could end with newlines.  The easiest
way to handle extra newlines is to add them to the grammar.
Probably, if we end up doing auto-formatting, we'll remove those
and account for it in the `top_parse` function.  That way we can
always have a consistent number of newlines at the end.