First try at converting Google Docstring.
A bunch of tests are broken, but it seems to be working.  The
grammar being converted to CNF makes it a little harder to reason
over, since there are more exceptions.

Probably, defining a query language to iterate over the tree would
help (or at least make it easier to make updates to the parser.)