Use annotations to identify key items.
Because we're translating from BNF to CNF, the parse tree ends up a
little messy, and is hard to predict ahead of time.  Obviously, we could
just rewrite the docstring class every time we change the grammar. (We'd
just have to have very thorough tests.)  To make the process easier, we
could generate all possible paths which could lead to a target token.

However, an easier approach is to use the annotations which were built
into the bnf_to_cnf converter as a means of handling invalid docstrings.
These annotations are more general (in fact, they're just a class.)
So, it's a perfect use case for simplifying the tree traversal.
(Since we can just record the annotations while traversing the tree.)

One downside to this is that some leaves are only comprehensible in the
context of their higher branch.  For example, noqas can target
particular arguments or exceptions for exclusion from errors.  They can
either specify them explicitly, or through context (by being placed on
the argument or exception in question.)  This means that those branches
end up being traversed twice.

We can probably write some careful conditions to avoid this unless the
user actually intends to use the context in this way.