Add returns grammar without type constraint.
We needed to have two separate grammars for the returns section:
one which specifies a type signature, and one which does not.  This
could also have been solved by adding probabilities to the grammar.  We
may end up doing this yet, but this approach (while less maintainable),
is much easier.