Add external imports to bnf_to_cnf
This will make the build process a little bit more sane, and allow us to
make more checks for validity in the future. (If we want to.)

External imports will always take the form:

    from <source> import (
        <filename>,
    )

Where there could be multiple filenames.  So, basically, it's the same
as the python from-import statement.  (Except that you have to have the
parentheses in this case.)