Add weight to sequences.
Now, when specifying a grammar, we can assign a weight as an
integer to a given sequence in a production.  For example, we could have
the following grammar:

    <item-head> ::= 99 <argument> <colon> <newline>
                  |  1 <argument> <newline>

    <argument> ::= <word>

Which would prefer the first production over the second.  The
weight is optional.  If there is not weight stated, then the
weight can be considered 0. (That is, a given weight will be
preferred over the default.)  This means that if we give a weight
to a production, then we should give a weight to any production
which can be used in its place.