Add optional newline to argument items.
The argument parsing wasn't working because it didn't have a newline
after each item (optional, of course.  The last line in the arguments
section would not have a newline, maybe.)

The same needs to be done for the other sections.

Now, the only problem is that the underindented line is consuming the
other arguments.  This is okay for the moment, but we could probably
prevent it by excluding newlines which start with a word and a colon,
so that it would force a new argument.  That may be more hassle than
it's worth, and will muddy up the grammar.

It's starting to get to the point where adding the normal EBNF features,
like `?` and `*` and `+` might be a good idea.