Add tool to begin integration testing.
With the new parser, regressions may have been introduced.  To try to
avoid regressions, this adds a utility to clone git repositories and
extract docstrings from them.

This also adds a utility to annotate the represented docstrings with
which arguments, exceptions, etc. they specify.  This should allow the
annotated docstrings to be used as goldens for regression testing.

Some features which need to be added before this all can work:
1) dragging words from the docstring to attributes/raises boxes,
2) representing which sections are present in the docstring.