Add argument checking to integration tests.
After looking, it appears that sphinx "param" items aren't supposed to
include the type.  (The type should appear separately.)  However, a lot
of sphinx comments combine the type.  So it's not entirely clear what
should happen.

Darglint will currently display an error for those, which is probably
not a bad thing.  However, the error may be confusing for users who
expect a type. (It will simply say that no argument matching "x y"
exists.)  So, we should add a style error and try to parse it correctly,
if possible.