Add preformance test to tox.
This creates a file, .performance_testrun, when run.  If the test
passes, it updates this file with the result.

A performance regression, in this case, means a 2 standard deviation
change in mean time to parse.  This may have to be adjusted.  For one,
the mean time to parse for small docstrings isn't much of a concern,
because it's very fast already.  Really, we're concerned about the
increase in time to parse outliers.  For another, it's unlikely that
such a large change will be introduced in one go.  More probably, we'll
introduce multiple small changes that increase over time.

One way of tackling this will be to curate the goldens a little bit
more.  We should have a lot of large docstrings for testing goldens, and
only a few short ones. (The short ones aren't very interesting, anyway.)