Add script to report grammar size for tox.
After trying to increase performance in some obvious ways
(using a more sparse data structure for CYK, adding threading to
the individual sections parsed in the docstring, and running a
profiler over the regression tests), there are no very simple
ways to improve performance.

The next step is to try to reduce the size of the grammar.
By reducing the grammar size, we should be able to increase
the overal speed of the parse.  We can likely reduce grammar
size by making sure that all productions are reachable in a
grammar.  Before attempting this, though, we need to know the
starting point. (Since it may require actually hand-editing the
grammar in some ways.)