Added recursive descent parser for long descr.
After looking a little closer at which docstrings took the most time,
it appeared that docstrings with a large long-description section took
substantially longer. This makes sense, as these are frequently the
longest sections, and can have the most tokens.  Also, long-description
is the backup for every failed section, so if an arguments section takes
a long time, the long description will also.

Introducing a recursive descent parsing algorithm for long-description
reduced the time to parse for the worst 1% of docstrings by half.  Now,
rather than 6 seconds, the longest docstring parse time is 2 seconds.

Of course, this slightly increases the complexity of parsing, but I
think it's worth the trade-off.