Move test/debug-only functions to utils file.
Most of the bigger tools for testing, developing, and debugging Darglint
are in the bin/ directory.  Some smaller functions remained on the
CykNode object.  Moving these functions to the utils module won't
significantly reduce the memory footprint of darglint, or increase start
times.  However, it will reduce the possibility of parse errors when
running darglint.