Handle arguments, other arguments for Numpy.
The numpy format now handles an arguments section and
an other arguments section.  There are a lot of details
which are still missing, which we'll have to cover in
unit tests later on.  For example, having extra blank
lines between items, or using the incorrect heading for
the arguments section. (For numpy, I believe it should
always be "Parameters" and "Other Parameters").

Broke a couple of tests for Google format -- it will have
to be fixed afterwards.  Mostly, it's just unhandled keywords
which are recognized now.