Finished initial grammar for Numpy.
The next steps will be to separate the grammar out into subsections,
split the docstring into sections and use a combinator.  Then, we'll
have to define a new Docstring type for the numpy format, a means of
selecting that format through the configuration file and command
line arguments.  Finally, we'll have to refine the definitions to
handle some incorrect versions, and implement the tests which were
passed on here.