Add goldens for Numpy.
I've added 100 goldens for numpy.  They've pointed out the main
problem with the numpy parser at the moment.  That is, the
`top_parse` function for its parser combinator doesn't correctly
split sections.  So, the last recognized section (normally "Returns")
gathers all the cruft at the bottom.  This is problematic because in
the case of large docstrings, it can take a very long time to parse.
The longest docstring in the numpy repository, for example, takes
close to 4 hours to parse.  (The last section contains over 1000 tokens).
We can partially solve this by adding the other sections into the
grammar, and parsing them with `long_description`.  Or, we could
enforce there being no line breaks in some sections.  But that is
likely being too strict.