Add "See Also", "Notes", and "Examples" sections.
Adding these other sections makes parsing (somewhat ironically)
faster for Numpy docstrings.  This way, the other sections are
bounded better.

A better fix will be to have a more intelligent means of splitting
the tokens in the `top_parse` function.  For the moment, though,
this works well for being able to just run tests a little more
efficiently.  If someone uses a non-standard section, it will
degrade quickly.