Handle multi-line body for sphinx item bodies.
Splits in item-bodies were not being handled correctly.
There's still one use-case which isn't handled correctly:
when an item has nothing immediately after the head, and
then has an indented block afterwards.

In order to handle this, however, we have to either accept
item bodies with no content (which is probably the best
option, as we should mark it as a formatting error), or
we rewrite the top-level parser.

Resolves: Issue #63.