Update bnf_to_cnf to handle elision to grammar.
In some cases, a node name is elided up to the grammar start symbol.
For example, in the following grammar:

    Grammar: ElidedSymbolGrammar

    start: <start>

    <start>
        ::= <a> <b>
        | <a>

    <a> ::= "0"
    <b> ::= "1"

Then the productions which accept the string "0" as a language
is elided up to the grammar (because both `<a>` and `<start>`
are single-element productions, which get replaced by whatever
they point to.

The fix is to identify such productions and skip them, as they
will already have been included when the start symbol is translated.

Note that any annotations will fall off -- just as with any single-
element production.  So you have to push the annotations down into
the chlidren.